(scala - GeoTrellis) Visualizing layers as a TIFF

+++ scala/geotrellis/src/main/scala/playground/Distance.scala
@@ -2,7 +2,6 @@ package playground
 
 import geotrellis.raster._
 import geotrellis.raster.render._
-import geotrellis.raster.render.ascii._
 import geotrellis.vector.{ Extent, Point }
 
 // --- //
@@ -11,8 +10,8 @@ object Distance {
 
   lazy val euclid: Tile = {
     val extent: RasterExtent = RasterExtent(Extent(0, 0, 250, 250), 250, 250)
-    val ps0: Array[Point] = Array.range(50, 200).map(n => Point(n, n))
-    val ps1: Array[Point] = Array.range(50, 200).map(n => Point(n, (-1 * n) + 250))
+    val ps0: Array[Point] = Array.range(50, 200).map(n => Point(n.toDouble, n.toDouble))
+    val ps1: Array[Point] = Array.range(50, 200).map(n => Point(n.toDouble, (-1 * n.toDouble) + 250))
 
     (ps0 ++ ps1).euclideanDistanceTile(extent)
   }
